refining project

+++ src/main/java/com/example/latihan/crud/repositories/ProductRepositories.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepositories extends JpaRepository<ProductEntity, Integer> {
+public interface ProductRepositories extends JpaRepository<ProductEntity, Long> {
 }
